Fix some tests & file names; make minimum size of code to be 3
