+++ src/Url.Shortener.Api/Domain/Url/Get/GetUrlExceptions.cs
@@ -4,5 +4,5 @@ namespace Url.Shortener.Api.Domain.Url.Get;
 
 internal static class GetUrlExceptions
 {
-    public static NotFoundException UrlNotFound() => new("Could not find url metadata for the specified code.");
+    public static NotFoundException CodeNotFound() => new("Could not find url metadata for the specified code.");
 }
