+++ src/Url.Shortener.Api/Domain/Url/Redirect/RedirectUrlExceptions.cs
@@ -4,5 +4,5 @@ namespace Url.Shortener.Api.Domain.Url.Redirect;
 
 internal static class RedirectUrlExceptions
 {
-    public static NotFoundException UrlNotFound() => new("Could not find matching url for the specified code.");
+    public static NotFoundException CodeNotFound() => new("Could not find matching url for the specified code.");
 }
