+++ src/Url.Shortener.Api/Domain/Url/RuleBuilderExtensions.cs
@@ -6,7 +6,7 @@ namespace Url.Shortener.Api.Domain.Url;
 internal static class RuleBuilderExtensions
 {
     public static IRuleBuilderOptions<T, string> EnsureValidCode<T>(this IRuleBuilder<T, string> ruleBuilder) =>
-        ruleBuilder.MinimumLength(5)
+        ruleBuilder.MinimumLength(3)
                    .MaximumLength(Constants.MaxCodeLength)
                    .Must(x => Uri.IsWellFormedUriString(x, UriKind.Relative))
                    .WithMessage("The '{PropertyName}' must be a valid relative uri.")
