+++ test/Url.Shortener.Api.UnitTests/Domain/Url/Create/CreateUrlRequestValidatorFixture/WhenValidatingRequest.cs
@@ -43,7 +43,7 @@ public class WhenValidatingRequest
             var fixture = new Fixture();
             var validUrl = $"https://{fixture.Create<string>()}.com/{fixture.Create<string>()}";
             Add(validUrl, default);
-            Add(validUrl, "abcde");                                              // 5 length string
+            Add(validUrl, "abc");                                              // 3 length string
             Add(validUrl, "tAJqPRr@ywgK5Av+nzu6c@WgB5AfxWj4Mmnt5o3UuqEqYFsARJ"); // 50 length string
         }
     }
