+++ test/Url.Shortener.Api.UnitTests/Domain/Url/Create/CreateUrlRequestValidatorFixture/WhenValidatingRequestWithInvalidCode.cs
@@ -54,7 +54,7 @@ public class WhenValidatingRequestWithInvalidCode
         {
             Add(string.Empty);
             Add("   ");
-            Add("tAJq");                                                // string with 4 length
+            Add("tA");                                                // string with 2 length
             Add("tAJqPRr@ywgK5Av*nzu6c@Wg+5#fxWj4Mmnt5o3UuqEqYFsARJT"); // string with 51 length
             Add("abcde/ab");                                            // path separator ('/') character not allowed
             Add("abcde\\ab");                                           // invalid relative uri
