+++ test/Url.Shortener.Api.UnitTests/Domain/Url/Create/ServiceCollectionExtensionsFixture/WhenBuildingServiceProvider.cs
@@ -9,13 +9,13 @@ namespace Url.Shortener.Api.UnitTests.Domain.Url.Create.ServiceCollectionExtensi
 
 public class WhenBuildingServiceProvider
 {
-    private readonly Action<CodeGeneratorOptions> _configureUrlShortenerOptions;
+    private readonly Action<CodeGeneratorOptions> _configureCodeGeneratorOptions;
     private readonly IServiceCollection _serviceCollection;
 
     public WhenBuildingServiceProvider()
     {
         var urlShortenerOptions = new UrlShortenerOptionsBuilder().Build();
-        _configureUrlShortenerOptions = options =>
+        _configureCodeGeneratorOptions = options =>
         {
             options.Characters = urlShortenerOptions.Characters;
             options.UrlSize = urlShortenerOptions.UrlSize;
@@ -48,6 +48,6 @@ public class WhenBuildingServiceProvider
         Assert.NotNull(provider.GetService<IOptions<CodeGeneratorOptions>>());
     }
 
-    private IServiceProvider WhenBuilding() => _serviceCollection.AddCreateUrlServices(_configureUrlShortenerOptions)
+    private IServiceProvider WhenBuilding() => _serviceCollection.AddCreateUrlServices(_configureCodeGeneratorOptions)
                                                                  .BuildServiceProvider();
 }
