+++ test/Url.Shortener.Api.UnitTests/Domain/Url/Get/GetUrlRequestHandlerFixture/WhenHandlingRequestAndUrlMetadataCannotBeFound.cs
@@ -53,7 +53,7 @@ public class WhenHandlingRequestAndUrlMetadataCannotBeFound
         var exception = await Record.ExceptionAsync(WhenHandlingAsync);
 
         var notFoundException = (exception as NotFoundException)!;
-        var expectedException = GetUrlExceptions.UrlNotFound();
+        var expectedException = GetUrlExceptions.CodeNotFound();
 
         Assert.Equal(expectedException.Message, notFoundException.Message);
     }
