+++ test/Url.Shortener.Api.UnitTests/Domain/Url/ServiceCollectionExtensionsFixture/WhenBuildingServiceProvider.cs
@@ -9,13 +9,13 @@ namespace Url.Shortener.Api.UnitTests.Domain.Url.ServiceCollectionExtensionsFixt
 
 public class WhenBuildingServiceProvider
 {
-    private readonly Action<CodeGeneratorOptions> _configureUrlShortenerOptions;
+    private readonly Action<CodeGeneratorOptions> _configureCodeGeneratorOptions;
     private readonly IServiceCollection _serviceCollection;
 
     public WhenBuildingServiceProvider()
     {
         var urlShortenerOptions = new UrlShortenerOptionsBuilder().Build();
-        _configureUrlShortenerOptions = options =>
+        _configureCodeGeneratorOptions = options =>
         {
             options.Characters = urlShortenerOptions.Characters;
             options.UrlSize = urlShortenerOptions.UrlSize;
@@ -32,6 +32,6 @@ public class WhenBuildingServiceProvider
         Assert.Null(exception);
     }
 
-    private IServiceProvider WhenBuilding() => _serviceCollection.AddUrlServices(_configureUrlShortenerOptions)
+    private IServiceProvider WhenBuilding() => _serviceCollection.AddUrlServices(_configureCodeGeneratorOptions)
                                                                  .BuildServiceProvider();
 }
